+++ gamechangerml/src/data_transfer/s3_download.py
@@ -8,7 +8,7 @@ from os import makedirs
 from os.path import join, exists, basename
 
 from gamechangerml.src.services.s3_service import S3Service
-from gamechangerml.utils import configure_logger
+from gamechangerml.src.utilities import configure_logger
 from gamechangerml.configs import S3Config
 from gamechangerml.api.utils import processmanager
 from gamechangerml.src.data_transfer import delete_local_corpus
