--- gamechangerml/utils/__init__.py
@@ -1,1 +1,0 @@
-from .logger import configure_logger
